requirements.txt
asgiref==3.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorclass==2.2.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
docopt==0.6.2
idna==3.2
importlib-metadata==4.6.0
iniconfig==1.1.1
more-itertools==8.8.0
packaging==20.9
pip-upgrader==1.4.15
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
requests==2.25.1
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.1
terminaltables==3.1.0
toml==0.10.2
typing==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
zipp==3.4.1
